requirements.txt
# Core Dependencies
streamlit==1.28.0
pandas==2.1.1
python-dotenv==1.0.0
requests==2.31.0
python-dateutil==2.8.2
pydantic==2.3.0
pydantic-settings==2.0.3

# API Clients
googlemaps==4.10.0
serpapi==0.1.5

# Data Processing
numpy==1.26.0
pytz==2023.3

# Caching
cachetools==5.3.1

# Utilities
tqdm==4.66.1
python-slugify==8.0.1

# Optional (for enhanced features)
geopy==2.4.0  # For geocoding
plotly==5.17.0  # For data visualization
